+++ geoopt/linalg/__init__.py
@@ -1,0 +1,1 @@
+from .batch_linalg import svd, qr, sym, extract_diag, matrix_rank
