+++ geoopt/linalg/batch_linalg.py
